requirements.txt
amqp==2.5.2
appdirs==1.4.4
asgiref==3.2.10
Authlib==0.14.1
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudinary==1.17.0
cryptography==2.9
defusedxml==0.6.0
dict2xml==1.7.0
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.11
django-cloudinary-storage==0.2.3
django-cors-headers==3.5.0
django-db-logger==0.1.9
django-debug-toolbar==3.1
django-elasticsearch-dsl==7.1.4
django-extensions==2.2.8
django-filter==2.2.0
django-heroku==0.3.1
django-online-users==0.3
django-rest-swagger==0.3.10
django-session-timeout==0.1.0
django-six==1.0.4
django-taggit==1.2.0
djangocric==1.0.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-xml==2.0.0
drf-tweaks==0.9.3
elasticsearch==7.5.1
elasticsearch-dsl==7.2.1
filelock==3.0.12
Flask==1.1.2
Flask-Login==0.5.0
future==0.18.2
gunicorn==19.9.0
heroku==0.1.4
idna==2.10
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mock==4.0.2
Pillow==7.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pysports==1.0.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3
raven==6.10.0
redis==2.10.3
requests==2.24.0
simple-salesforce==1.0.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.26
Werkzeug==1.0.1
whitenoise==4.1.2
zipp==3.1.0
